backend/worker-python/scraper/requirements.txt
# Database
psycopg2-binary==2.9.9

# Scraping
playwright==1.57.0
requests==2.31.0
beautifulsoup4==4.12.3

# Utils
python-dotenv==1.0.1
